move test to cleaning package
