+++ plume-file-core/src/test/java/com/coreoz/plume/file/cleaning/FileExtensionCleaningTest.java
@@ -1,6 +1,5 @@
-package com.coreoz.plume.file.utils;
+package com.coreoz.plume.file.cleaning;
 
-import com.coreoz.plume.file.cleaning.FileExtensionCleaning;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
